renamed scripts to autogpt
