+++ README.md
@@ -119,7 +119,7 @@ pip install -r requirements.txt
    _(Type this into your CMD window)_
 
 ```
-python scripts/main.py
+python autogpt/main.py
 ```
 
 2. After each of AUTO-GPT's actions, type "NEXT COMMAND" to authorise them to continue.
@@ -132,7 +132,7 @@ You will find activity and error logs in the folder `./output/logs`
 To output debug logs:
 
 ```
-python scripts/main.py --debug
+python autogpt/main.py --debug
 ```
 
 ## 🗣️ Speech Mode
@@ -140,7 +140,7 @@ python scripts/main.py --debug
 Use this to use TTS for Auto-GPT
 
 ```
-python scripts/main.py --speak
+python autogpt/main.py --speak
 ```
 
 ## 🔍 Google API Keys Configuration
@@ -269,7 +269,7 @@ Use at your own risk.
 1. Run the `main.py` Python script in your terminal:
 
 ```
-python scripts/main.py --continuous
+python autogpt/main.py --continuous
 
 ```
 
@@ -280,7 +280,7 @@ python scripts/main.py --continuous
 If you don't have access to the GPT4 api, this mode will allow you to use Auto-GPT!
 
 ```
-python scripts/main.py --gpt3only
+python autogpt/main.py --gpt3only
 ```
 
 It is recommended to use a virtual machine for tasks that require high security measures to prevent any potential harm to the main computer's system and data.
@@ -353,8 +353,8 @@ This project uses [flake8](https://flake8.pycqa.org/en/latest/) for linting. We 
 To run the linter, run the following command:
 
 ```
-flake8 scripts/ tests/
+flake8 autogpt/ tests/
 
 # Or, if you want to run flake8 with the same configuration as the CI:
-flake8 scripts/ tests/ --select E303,W293,W291,W292,E305,E231,E302
+flake8 autogpt/ tests/ --select E303,W293,W291,W292,E305,E231,E302
 ```
