+++ autogpt/image_gen.py
